GH-1920: keep commonJS in tailwind css
this close GH-1920
