+++ packages/tailwindcss-config/tailwind.config.cjs
@@ -1,0 +1,12 @@
+const tailwindCssFormsPlugin = require('@tailwindcss/forms');
+
+const tailwindConfig = {
+  plugins: [
+    tailwindCssFormsPlugin({
+      strategy: 'class',
+    }),
+  ],
+  prefix: 'tw-',
+};
+
+module.exports = tailwindConfig;
